cap layer names at 45 characters
When loading into postgis, the default limit on table/index names is 64
characters. I think this is a compile time setting, but 64 seems to be
the common one used. We prefix a "spatial_" and suffix a "_geometry" to
the name, so we only get 47 chars to work with. And reserving 2 chars
for generating unique names by appending integers leaves us with 45.