delegate to GeoServerExtensions.getProperty
Instead of manually checking environment variables, delegate to the
GeoServerExtensions.getProperty utility method to do so. This changes
the lookup order, but is now consistent with other geoserver lookups.
We also don't know the source of the override, just whether an override
exists or not. Update the interface and UI accordingly.