Upgrade json-schema-validator to support building with maven 3
Upgrading the org.kitchen-eel:json-schema-validator version fixes
the build with maven 3, due to a known bug[1] that's fixed in newer
versions. Note that the org.kitchen-eel namespace has moved to
com.github.fge[2].

This also requires an upgrade of the guava version declared in
the root pom.xml, as required by json-schema-validator for which
guava is a dependency too. We're upgrading to 14.0.1 although
json-schema-validator requires 13.0.1 to also support the same
version that geogit uses.

[1] <https://github.com/fge/json-schema-validator/issues/39>
[2] <https://groups.google.com/forum/#!msg/json-schema-validator/V_lAFBOuTAY/dIJgC6lk_mgJ>